src/requirements.txt
Package           Version
----------------- --------
annotated-types   0.7.0
anyio             4.9.0
asyncpg           0.30.0
bcrypt            4.3.0
cffi              1.17.1
click             8.2.1
cryptography      45.0.3
ecdsa             0.19.1
fastapi           0.115.12
h11               0.16.0
idna              3.10
pip               24.3.1
pyasn1            0.4.8
pycparser         2.22
pydantic          2.11.5
pydantic_core     2.33.2
python-dotenv     1.1.0
python-jose       3.4.0
rsa               4.9.1
six               1.17.0
sniffio           1.3.1
starlette         0.46.2
typing_extensions 4.13.2
typing-inspection 0.4.1
uvicorn           0.34.2
